+++ blessed/win_terminal.py
@@ -104,7 +104,7 @@ def _win32_resize_to_seq(fd: int) -> str:
     return f'\x1b[48;{size.lines};{size.columns};0;0t'
 
 
-class Terminal(_Terminal):
+class Terminal(_Terminal):  # pylint: disable=attribute-defined-outside-init
     """Windows subclass of :class:`Terminal`."""
 
     def __init__(self,
@@ -423,7 +423,7 @@ class Terminal(_Terminal):
             self._event_buf.clear()
             del self._dec_mode_cache[
                 _DecPrivateMode.IN_BAND_WINDOW_RESIZE]
-            self._preferred_size_cache = None
+            self._preferred_size_cache = None  # pylint: disable=attribute-defined-outside-init
             win32.set_console_mode(filehandle, save_mode)
 
     def does_inband_resize(self, timeout: float = 1.0) -> bool: